remove avp kernel module source code
The AVP kernel module source code is now stored in its own public repo.

    https://github.com/Wind-River/titanium-cloud-avp-kmod

The AVP PMD is retained for archiving purposes to support customers that need
to use an older DPDK version but it will eventually be discontinued in favor of
the version included within the DPDK itself.  We recommend moving all
applications forward to the latest stable DPDK version available at:

    https://dpdk.org